chore(rust/chunk): move `Chunk` to `rolldown_common`
