--- crates/rolldown/src/chunk/types/mod.rs
@@ -1,1 +1,0 @@
-pub mod pre_renderer_chunk;
