+++ crates/rolldown/src/chunk_graph.rs
@@ -1,10 +1,10 @@
 use index_vec::IndexVec;
 use rolldown_common::{ChunkId, NormalModuleId};
 
-use super::chunk::ChunksVec;
+use crate::type_alias::IndexChunks;
 
 #[derive(Debug)]
 pub struct ChunkGraph {
-  pub chunks: ChunksVec,
+  pub chunks: IndexChunks,
   pub module_to_chunk: IndexVec<NormalModuleId, Option<ChunkId>>,
 }
