+++ crates/rolldown/src/type_alias.rs
@@ -1,0 +1,4 @@
+use index_vec::IndexVec;
+use rolldown_common::{Chunk, ChunkId};
+
+pub type IndexChunks = IndexVec<ChunkId, Chunk>;
