+++ crates/rolldown_common/src/chunk/mod.rs
@@ -1,23 +1,15 @@
 // cSpell:disable
-use index_vec::IndexVec;
-use rolldown_common::{ChunkId, FileNameTemplate, NormalizedBundlerOptions};
+use crate::{
+  ChunkId, ChunkKind, ExternalModuleId, FileNameTemplate, NamedImport, NormalModuleId,
+  NormalizedBundlerOptions, SymbolRef,
+};
 pub mod types;
 
-pub type ChunksVec = IndexVec<ChunkId, Chunk>;
-
-use rolldown_common::{
-  ChunkKind, ExternalModuleId, NamedImport, NormalModuleId, RenderedChunk, Specifier, SymbolRef,
-};
 use rolldown_rstr::Rstr;
-use rolldown_sourcemap::SourceMap;
 use rolldown_utils::BitSet;
 use rustc_hash::FxHashMap;
 
-#[derive(Debug, Clone)]
-pub struct CrossChunkImportItem {
-  pub export_alias: Option<Specifier>,
-  pub import_ref: SymbolRef,
-}
+use self::types::cross_chunk_import_item::CrossChunkImportItem;
 
 #[derive(Debug, Default)]
 pub struct Chunk {
@@ -31,12 +23,6 @@ pub struct Chunk {
   pub imports_from_external_modules: FxHashMap<ExternalModuleId, Vec<NamedImport>>,
   // meaningless if the chunk is an entrypoint
   pub exports_to_other_chunks: FxHashMap<SymbolRef, Rstr>,
-}
-
-pub struct ChunkRenderReturn {
-  pub code: String,
-  pub map: Option<SourceMap>,
-  pub rendered_chunk: RenderedChunk,
 }
 
 impl Chunk {
