+++ crates/rolldown_common/src/chunk/types/cross_chunk_import_item.rs
@@ -1,0 +1,7 @@
+use crate::{Specifier, SymbolRef};
+
+#[derive(Debug, Clone)]
+pub struct CrossChunkImportItem {
+  pub export_alias: Option<Specifier>,
+  pub import_ref: SymbolRef,
+}
