+++ crates/rolldown_common/src/chunk/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod cross_chunk_import_item;
+pub mod pre_renderer_chunk;
