+++ crates/rolldown_common/src/chunk/types/pre_renderer_chunk.rs
